requirements.txt
Flask>=2.2.0
openai>=1.0.0
gunicorn
python-dotenv
